fix: resolve backend-size CI failures after main rebase
Move factory-dir resolution helpers and layout CLI tests into dedicated
files, and split the stale-layout pruning functional test so backend-size
checks pass on the rebased head.
